refactor(cli): pull SubgraphPaginationArgs into common crate
